+++ R/vtype.R
@@ -8,7 +8,8 @@
 #' @examples
 #' vtype(dst_beta(1, 2))
 #' vtype(dst_bern(0.4))
-#' vtype(distribution())
+#' vtype(distribution(cdf = pnorm, density = dnorm,
+#'                    .support = continuous()))
 #' @export
 vtype <- function(distribution) {
   checkmate::assert_class(distribution, "dst")